fix: renama getPlacement file
