+++ packages/design-system/tooltip/index.tsx
@@ -4,7 +4,7 @@ import { AccessibilityInfo, Pressable, StyleSheet, View } from "react-native";
 import { AnimatePresence, MotiView } from "moti";
 
 import { tw as tailwind } from "../tailwind";
-import { Placement, PlatformRect } from "./getPlacement";
+import { Placement, PlatformRect } from "./get-placement";
 import { Position } from "./position";
 import { TooltipContent } from "./tooltop.content";
 import { TooltipProps } from "./types";
